meta-agl-bsp: refresh mesa virgl patch
Refresh the mesa patch applied dynamically when meta-virtualization is
being used so it applies cleanly.  It is still unknown if the patch is
required for the KVM demo, but fixing it up is required to have the
IC EG container demos to build.

Bug-AGL: SPEC-5123

Change-Id: Ib1c8b623d78ca0453d3418e565dbe86bdc344fb5
Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/29898
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account
ci-image-boot-test: Jenkins Job builder account