Update machine local.conf fragment numbering
Lower numbering of the local.conf fragments in the machine templates
so they will be guaranteed to be before feature added fragments as
opposed to mixed in with them.  This allows referring to MACHINE in
feature fragments in a stable fashion, and makes the generated
local.conf a bit more readable.

Bug-AGL: SPEC-5306

Change-Id: I193027dfeadda20fee1f74f925d980378ea076f5
Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/30545
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account
ci-image-boot-test: Jenkins Job builder account