agl-compositor: Bump SRCREV to pick-up latest changes
This introduces a bunch of new features to the AGL compositor.
Specifcally, the ability the load multiple backends at the same, and we
have added support for PipeWire back-end. This allows streaming of
output with Pipewire and using VNC/RDP for relaying input back
the compositor, a functionality we had previously with Waltham.

Some other changes is that we should be ready to use libweston 14 when
that arrives.

With this release we've now removed agl-shell-desktop private extension
in favor of using gRPC proxy API. We also removed remoting plug-in in
favour of using PipeWire for streaming.

In more detail the following changes have been added, since the previous
update:

- 896f27c meson.build: Update meson version to 1.0
- 76a2aad clients/stream-pipewire-output: Add script to stream PipeWire output
- 57099d3 layout: Move the logging and activation at first mapping
- c8f2664 compositor: Add support for loading multiple back-ends
- de6dfd2 compositor: Add support for loading PipeWire backend
- 2487a38 compositor: Remove loading of remoting plug-in
- c023bf5 desktop: Remove notification for mapping first time
- f2ebad3 compositor: Remove agl-shell-desktop extension
- d60ccb8 layout: Use weston_view_move_to_layer when adding to hidden layer
- bb22f78 layout: Use weston_view_move_to_layer helper for de-activating views
- 6d88952 layout: Use weston_view_move_layer for all other surface roles
- c48b210 layout: Update layout to use weston_view_move_to helper
- b6f811c layout: Update layout to use weston_view_move_to helper

Change-Id: I90975e6e194b757507f8bc7433ab6e9debc3a8d2
Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/30595
ci-image-boot-test: Jenkins Job builder account
ci-image-build: Jenkins Job builder account
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Tested-by: Jenkins Job builder account