weston_aglcore.inc: Always build the PipeWire backend
We need the PipeWire backend in agl-kvm-demo but it might be useful in
other builds as well, so let's build it by default. Other users can just
include the package when they require it.

Furthermore, with this change we remove weston-remoting and with it we
no longer build remoting plug-in when agl-weston-remoting feature is
supplied.

Bug-AGL: SPEC-5235
Change-Id: I140c46aa33ce832100f4e696aa3bcfef6d43522a
Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/30611
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Tested-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>